requirements.txt
locust==2.15.0
Faker==17.6.0
mysql-connector-python==8.0.32
python-dotenv==1.0.0
